Fix add device profiles

+++ HES.Core/Services/DeviceService.cs
@@ -1,4 +1,5 @@
 ﻿using HES.Core.Entities;
+using HES.Core.Enums;
 using HES.Core.Interfaces;
 using HES.Core.Models;
 using Hideez.SDK.Communication;
@@ -369,6 +370,12 @@ namespace HES.Core.Services
                 throw new ArgumentNullException(nameof(profileId));
             }
 
+            var state = await _deviceRepository.Query().Where(x => devicesId.Contains(x.Id)).Select(d => d.State != DeviceState.OK).AnyAsync();
+            if (state)
+            {
+                throw new Exception("You have chosen a device with a status that does not allow changing the profile.");
+            }
+
             var profile = await _deviceAccessProfileRepository.GetByIdAsync(profileId);
             if (profile == null)
             {
@@ -383,7 +390,7 @@ namespace HES.Core.Services
                     device.AcceessProfileId = profileId;
                     await _deviceRepository.UpdateOnlyPropAsync(device, new string[] { "AcceessProfileId" });
 
-                    if (device.MasterPassword != null)
+                    if (device.MasterPassword != null && device.EmployeeId != null)
                     {
                         // Delete all previous tasks for update profile
                         await _deviceTaskService.RemoveAllProfileTasksAsync(device.Id);
